# Improve minimize performance