@wlammen 2022-07-17
> https://github.com/eteran/c-vector
And it is C89 (ANSI C https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf) compatible.  I wonder whether we should give up on this and at least advance to C11.