@david-a-wheeler 2022-07-17
> I wonder whether we should give up on this and at least advance to C11.

We could move to C11 if that's helpful. The key is compiler support in deployed systems. [GCC seems to have good C11 support](https://gcc.gnu.org/wiki/C11Status#:~:text=GCC%204.9%20Changes%3A%20%E2%80%9CISO%20C11,optional%20C99%20features%20from%20Annexes).

However, I don't think that helps. C11 doesn't have a vector type as far as I know, and <https://github.com/eteran/c-vector> doesn't require C11. We should write code that is *compatible* with C11, but I think we should only require C11 if there's advantage to it. I didn't see any advantages in [this list of C11 additions](https://en.wikipedia.org/wiki/C11_%28C_standard_revision%29)