@wlammen 2022-07-17
> However, I don't think that helps.

See for example https://en.cppreference.com/w/c/io/fprintf to see how much C has developed since 1989.  Functions like snprintf really make sense. Or // style comments.

`gcc -O3 -funroll-loops -finline-functions -fomit-frame-pointer -DINLINE=inline -g  -Wall -Wextra   -o metamath metamath.o mmcmdl.o mmcmds.o mmdata.o mmfatl.o mmhlpa.o mmhlpb.o mminou.o mmpars.o mmpfas.o mmunif.o mmveri.o mmvstr.o mmword.o mmwtex.o`

This is the command executed by the build.sh script.  It does not check for ANSI C.  The option `-ansi` should be added to enforce this standard.  By the way, this can be done by changing line 4567 to new_CFLAGS="-Wall -Wextra -ansi".