@david-a-wheeler 2022-07-17
I agree snprintf and // are useful, but I those are C99 additions, not C11.

I'm not opposed to switching to a later C spec, I just want to make sure that a spec move is justified *and* that it is widely implemented by commonly-available & distributed compilers.

We should at *least* move up to C99, no question in my mind!!! C99 is very widely supported & I think we're already using its facilities. The only question is if we should move to C11 or not.