@digama0 2022-07-18
We are already on C11. The build script uses it and `//` can be spotted in the source code.