@savask 2022-07-18
>It's certainly possible to add one.

It is, but is it worth doing so for one application? Also, metamath-exe uses its own memory management system, for example, in code I used `poolMalloc` instead of `malloc` and `poolFree` instead of `free`. I was wondering if `nmbrString` can be used as vector of numbers, but I haven't been able to find a `push_back` equivalent in mmdata.c.