@wlammen 2022-07-18
> We are already on C99.

Indeed: CONTRIBUTIONS.md states this literally.